+++ examples/live/interactive_brokers_v2/_common.py
@@ -201,11 +201,11 @@ def build_ib_live_node(
         None,
         ib.InteractiveBrokersDataClientFactory(),
         ib.InteractiveBrokersDataClientConfig(
-            ibg_host=host,
-            ibg_port=port,
-            ibg_client_id=data_client_id,
+            host=host,
+            port=port,
+            client_id=data_client_id,
             connection_timeout=env_int("IB_V2_CONNECTION_TIMEOUT", 10),
-            request_timeout_secs=env_int("IB_V2_REQUEST_TIMEOUT", 30),
+            request_timeout=env_int("IB_V2_REQUEST_TIMEOUT", 30),
             market_data_type=ib.MarketDataType.DELAYED_FROZEN,
             instrument_provider=provider_config,
         ),
@@ -216,12 +216,12 @@ def build_ib_live_node(
             None,
             ib.InteractiveBrokersExecutionClientFactory(trader, ib_account_id(account_id)),
             ib.InteractiveBrokersExecClientConfig(
-                ibg_host=host,
-                ibg_port=port,
-                ibg_client_id=exec_client_id or data_client_id,
+                host=host,
+                port=port,
+                client_id=exec_client_id or data_client_id,
                 account_id=account_id,
                 connection_timeout=env_int("IB_V2_CONNECTION_TIMEOUT", 10),
-                request_timeout_secs=env_int("IB_V2_REQUEST_TIMEOUT", 30),
+                request_timeout=env_int("IB_V2_REQUEST_TIMEOUT", 30),
                 fetch_all_open_orders=False,
                 instrument_provider=provider_config,
             ),
