+++ examples/live/interactive_brokers_v2/with_databento_client.py
@@ -73,12 +73,12 @@ def main() -> None:
             None,
             ib.InteractiveBrokersExecutionClientFactory(trader_id, ib_account_id(account_id)),
             ib.InteractiveBrokersExecClientConfig(
-                ibg_host=host,
-                ibg_port=port,
-                ibg_client_id=env_int("IB_V2_EXEC_CLIENT_ID", 1312),
+                host=host,
+                port=port,
+                client_id=env_int("IB_V2_EXEC_CLIENT_ID", 1312),
                 account_id=account_id,
                 connection_timeout=env_int("IB_V2_CONNECTION_TIMEOUT", 10),
-                request_timeout_secs=env_int("IB_V2_REQUEST_TIMEOUT", 30),
+                request_timeout=env_int("IB_V2_REQUEST_TIMEOUT", 30),
                 fetch_all_open_orders=False,
                 instrument_provider=provider_config,
             ),
